+++ maven-plugin/src/main/java/com/aeontronix/enhancedmule/tools/DeployMojo.java
@@ -137,7 +137,7 @@ public class DeployMojo extends LegacyDeployMojo {
                     properties = findPrefixProperties(properties, ANYPOINT_DEPLOY_PROPERTIES);
                     final RuntimeDeploymentRequest request = new RuntimeDeploymentRequest(filename != null ? filename :
                             source.getFileName(), appName, source.getArtifactId(), buildNumber, vars, properties, propertyfile,
-                            target, getEnvironment(), injectEnvInfo, skipWait, skipProvisioning);
+                            target, getEnvironment(), injectEnvInfo, skipWait, skipProvisioning, getLegacyAppDescriptor());
                     request.setFileProperties(fileProperties);
                     request.setFilePropertiesPath(filePropertiesPath);
                     request.setFilePropertiesSecure(filePropertiesSecure);